+++ tests/generate_app_id_mapping.py
@@ -8,11 +8,10 @@ https://opensource.org/licenses/MIT.
 
 from oaaclient.templates import CustomApplication, Tag, OAAPermission, OAAPropertyType
 
+def generate_app_id_mapping():
+    """ generates a complete OAA custom application payload where local users, groups and roles are mapped by a unique identifier"""
 
-def generate_app():
-    """ generates a complete OAA custom application payload """
-
-    app = CustomApplication(name="pytest generated app", application_type="pytest", description="This is a test")
+    app = CustomApplication(name="pytest unique id app", application_type="pytest", description="This is a test")
 
     app.property_definitions.define_application_property("version", OAAPropertyType.STRING)
     app.set_property("version", "2022.2.2")
@@ -41,24 +40,28 @@ def generate_app():
                                          OAAPermission.MetadataRead,
                                          OAAPermission.MetadataWrite,
                                          OAAPermission.NonData
-                                        ]
+                                         ]
                               )
     app.add_custom_permission("View", [OAAPermission.DataRead,
                                        OAAPermission.MetadataRead,
-                                      ]
+                                       ]
                               )
 
     # define one property of every type
     app.property_definitions.define_local_user_property("is_guest", OAAPropertyType.BOOLEAN)
-    app.property_definitions.define_local_user_property("user_id", OAAPropertyType.NUMBER)
-    app.property_definitions.define_local_user_property("NAME", OAAPropertyType.STRING)
-    app.property_definitions.define_local_user_property("peers", OAAPropertyType.STRING_LIST)
     app.property_definitions.define_local_user_property("birthday", OAAPropertyType.TIMESTAMP)
 
-    username_list = ["bob", "marry", "sue", "rob"]
-    for username in username_list:
-        user = app.add_local_user(username)
-        user.add_identity(f"{username}@example.com")
+    user_list = [{"name": "Megan", "id": 1234},
+                 {"name": "Scott", "id": 1235},
+                 {"name": "Amanda", "id": 1236},
+                 {"name": "Paul", "id": 1237}
+                 ]
+
+    for user_info in user_list:
+        user_name = user_info["name"]
+        user_id = user_info["id"]
+        user = app.add_local_user(name=user_name, unique_id=user_id)
+        user.add_identity(f"{user_name}@example.com")
         # set all the properties to something
         user.is_active = True
         user.created_at = "2001-01-01T00:00:00.000Z"
@@ -66,41 +69,38 @@ def generate_app():
         user.deactivated_at = "2003-03-01T00:00:00.000Z"
         user.password_last_changed_at = "2004-04-01T00:00:00.000Z"
         user.set_property("is_guest", False)
-        user.set_property("user_id", username_list.index(username))
-        user.set_property("NAME", username.swapcase())
-        user.set_property("peers", username_list)
         user.set_property("birthday", "2000-01-01T00:00:00.000Z")
 
     # groups
     app.property_definitions.define_local_group_property("group_id", OAAPropertyType.NUMBER)
-    group1 = app.add_local_group("group1")
+    group1 = app.add_local_group("group1", unique_id="g1")
     group1.created_ad = "2001-01-01T00:00:00.000Z"
     group1.set_property("group_id", 1)
 
-    app.local_users["bob"].add_group("group1")
-    app.local_users["marry"].add_group("group1")
+    app.local_users[1234].add_group("g1")
+    app.local_users[1235].add_group("g1")
 
-    group2 = app.add_local_group("group2")
+    group2 = app.add_local_group("group2", unique_id="g2")
     group2.created_ad = "2001-01-01T00:00:00.000Z"
     group2.set_property("group_id", 2)
 
-    app.local_users["bob"].add_group("group2")
-    app.local_users["marry"].add_group("group2")
+    app.local_users[1235].add_group("g2")
+    app.local_users[1237].add_group("g2")
 
-    group3 = app.add_local_group("group3")
-    group3.add_group("group1")
-    group3.add_group("group2")
-    app.local_users["rob"].add_group("group3")
+    group3 = app.add_local_group("group3", unique_id="g3")
+    group3.add_group("g1")
+    group3.add_group("g2")
+    app.local_users[1236].add_group("g3")
 
     # roles
     app.property_definitions.define_local_role_property("role_id", OAAPropertyType.NUMBER)
     app.property_definitions.define_local_role_property("custom", OAAPropertyType.BOOLEAN)
 
-    role1 = app.add_local_role("role1", ["all", "Admin"])
+    role1 = app.add_local_role("role1", unique_id="r1",  permissions=["all", "Admin"])
     role1.set_property("role_id", 1)
     role1.set_property("custom", False)
 
-    role2 = app.add_local_role("role2", ["view"])
+    role2 = app.add_local_role("role2", unique_id="r2", permissions=["view"])
     role2.set_property("role_id", 1)
     role1.set_property("custom", True)
 
@@ -129,16 +129,16 @@ def generate_app():
     cog1.add_resource_connection("service_account@some-project.iam.gserviceaccount.com", "GoogleCloudServiceAccount")
 
     # authorizations
-    app.local_users["bob"].add_role("role1", apply_to_application=True)
-    app.local_groups["group2"].add_role("role2", resources=[thing1])
-    app.local_users["marry"].add_permission("view", resources=[thing2, cog1])
-    app.local_users["rob"].add_permission("manage", resources=[thing1], apply_to_application=True)
+    app.local_users[1235].add_role("r1", apply_to_application=True)
+    app.local_groups["g2"].add_role("r2", resources=[thing1])
+    app.local_users[1236].add_permission("view", resources=[thing2, cog1])
+    app.local_users[1237].add_permission("manage", resources=[thing1], apply_to_application=True)
 
     return app
 
 
 # Full App payload as string
-GENERATED_APP_PAYLOAD = """
+GENERATED_APP_ID_MAPPINGS_PAYLOAD = """
 {
   "custom_property_definition": {
     "applications": [
@@ -149,9 +149,6 @@ GENERATED_APP_PAYLOAD = """
         },
         "local_user_properties": {
           "is_guest": "BOOLEAN",
-          "user_id": "NUMBER",
-          "NAME": "STRING",
-          "peers": "STRING_LIST",
           "birthday": "TIMESTAMP"
         },
         "local_group_properties": {
@@ -178,18 +175,17 @@ GENERATED_APP_PAYLOAD = """
   },
   "applications": [
     {
-      "name": "pytest generated app",
+      "name": "pytest unique id app",
       "application_type": "pytest",
       "description": "This is a test",
       "local_users": [
         {
-          "name": "bob",
+          "name": "Megan",
           "identities": [
-            "bob@example.com"
+            "Megan@example.com"
           ],
           "groups": [
-            "group1",
-            "group2"
+            "g1"
           ],
           "is_active": true,
           "created_at": "2001-01-01T00:00:00.000Z",
@@ -198,25 +194,18 @@ GENERATED_APP_PAYLOAD = """
           "password_last_changed_at": "2004-04-01T00:00:00.000Z",
           "custom_properties": {
             "is_guest": false,
-            "user_id": 0,
-            "NAME": "BOB",
-            "peers": [
-              "bob",
-              "marry",
-              "sue",
-              "rob"
-            ],
             "birthday": "2000-01-01T00:00:00.000Z"
-          }
+          },
+          "id": "1234"
         },
         {
-          "name": "marry",
+          "name": "Scott",
           "identities": [
-            "marry@example.com"
+            "Scott@example.com"
           ],
           "groups": [
-            "group1",
-            "group2"
+            "g1",
+            "g2"
           ],
           "is_active": true,
           "created_at": "2001-01-01T00:00:00.000Z",
@@ -225,47 +214,17 @@ GENERATED_APP_PAYLOAD = """
           "password_last_changed_at": "2004-04-01T00:00:00.000Z",
           "custom_properties": {
             "is_guest": false,
-            "user_id": 1,
-            "NAME": "MARRY",
-            "peers": [
-              "bob",
-              "marry",
-              "sue",
-              "rob"
-            ],
             "birthday": "2000-01-01T00:00:00.000Z"
-          }
+          },
+          "id": "1235"
         },
         {
-          "name": "sue",
+          "name": "Amanda",
           "identities": [
-            "sue@example.com"
-          ],
-          "is_active": true,
-          "created_at": "2001-01-01T00:00:00.000Z",
-          "last_login_at": "2002-02-01T00:00:00.000Z",
-          "deactivated_at": "2003-03-01T00:00:00.000Z",
-          "password_last_changed_at": "2004-04-01T00:00:00.000Z",
-          "custom_properties": {
-            "is_guest": false,
-            "user_id": 2,
-            "NAME": "SUE",
-            "peers": [
-              "bob",
-              "marry",
-              "sue",
-              "rob"
-            ],
-            "birthday": "2000-01-01T00:00:00.000Z"
-          }
-        },
-        {
-          "name": "rob",
-          "identities": [
-            "rob@example.com"
+            "Amanda@example.com"
           ],
           "groups": [
-            "group3"
+            "g3"
           ],
           "is_active": true,
           "created_at": "2001-01-01T00:00:00.000Z",
@@ -274,16 +233,28 @@ GENERATED_APP_PAYLOAD = """
           "password_last_changed_at": "2004-04-01T00:00:00.000Z",
           "custom_properties": {
             "is_guest": false,
-            "user_id": 3,
-            "NAME": "ROB",
-            "peers": [
-              "bob",
-              "marry",
-              "sue",
-              "rob"
-            ],
             "birthday": "2000-01-01T00:00:00.000Z"
-          }
+          },
+          "id": "1236"
+        },
+        {
+          "name": "Paul",
+          "identities": [
+            "Paul@example.com"
+          ],
+          "groups": [
+            "g2"
+          ],
+          "is_active": true,
+          "created_at": "2001-01-01T00:00:00.000Z",
+          "last_login_at": "2002-02-01T00:00:00.000Z",
+          "deactivated_at": "2003-03-01T00:00:00.000Z",
+          "password_last_changed_at": "2004-04-01T00:00:00.000Z",
+          "custom_properties": {
+            "is_guest": false,
+            "birthday": "2000-01-01T00:00:00.000Z"
+          },
+          "id": "1237"
         }
       ],
       "local_groups": [
@@ -291,20 +262,23 @@ GENERATED_APP_PAYLOAD = """
           "name": "group1",
           "custom_properties": {
             "group_id": 1
-          }
+          },
+          "id": "g1"
         },
         {
           "name": "group2",
           "custom_properties": {
             "group_id": 2
-          }
+          },
+          "id": "g2"
         },
         {
           "name": "group3",
           "groups": [
-            "group1",
-            "group2"
-          ]
+            "g1",
+            "g2"
+          ],
+          "id": "g3"
         }
       ],
       "local_roles": [
@@ -318,7 +292,8 @@ GENERATED_APP_PAYLOAD = """
           "custom_properties": {
             "role_id": 1,
             "custom": true
-          }
+          },
+          "id": "r1"
         },
         {
           "name": "role2",
@@ -328,7 +303,8 @@ GENERATED_APP_PAYLOAD = """
           "tags": [],
           "custom_properties": {
             "role_id": 1
-          }
+          },
+          "id": "r2"
         }
       ],
       "tags": [],
@@ -429,23 +405,27 @@ GENERATED_APP_PAYLOAD = """
   ],
   "identity_to_permissions": [
     {
-      "identity": "bob",
+      "identity": "1234",
+      "identity_type": "local_user"
+    },
+    {
+      "identity": "1235",
       "identity_type": "local_user",
       "role_assignments": [
         {
-          "application": "pytest generated app",
-          "role": "role1",
+          "application": "pytest unique id app",
+          "role": "r1",
           "apply_to_application": true,
           "resources": []
         }
       ]
     },
     {
-      "identity": "marry",
+      "identity": "1236",
       "identity_type": "local_user",
       "application_permissions": [
         {
-          "application": "pytest generated app",
+          "application": "pytest unique id app",
           "resources": [
             "thing2",
             "thing2.cog1"
@@ -455,20 +435,16 @@ GENERATED_APP_PAYLOAD = """
       ]
     },
     {
-      "identity": "sue",
-      "identity_type": "local_user"
-    },
-    {
-      "identity": "rob",
+      "identity": "1237",
       "identity_type": "local_user",
       "application_permissions": [
         {
-          "application": "pytest generated app",
+          "application": "pytest unique id app",
           "permission": "manage",
           "apply_to_application": true
         },
         {
-          "application": "pytest generated app",
+          "application": "pytest unique id app",
           "resources": [
             "thing1"
           ],
@@ -477,16 +453,16 @@ GENERATED_APP_PAYLOAD = """
       ]
     },
     {
-      "identity": "group1",
+      "identity": "g1",
       "identity_type": "local_group"
     },
     {
-      "identity": "group2",
+      "identity": "g2",
       "identity_type": "local_group",
       "role_assignments": [
         {
-          "application": "pytest generated app",
-          "role": "role2",
+          "application": "pytest unique id app",
+          "role": "r2",
           "apply_to_application": false,
           "resources": [
             "thing1"
@@ -495,7 +471,7 @@ GENERATED_APP_PAYLOAD = """
       ]
     },
     {
-      "identity": "group3",
+      "identity": "g3",
       "identity_type": "local_group"
     }
   ]